Move enums to enums.py
* closes #122
